python_worker/requirements.txt
ultralytics>=8.3.0
opencv-python
httpx
python-dotenv
numpy
